Change language-change buttons to links
In the Django-based site, this was a view that one posts
to, and changes the user's selected language; which means that
every page will then be presented in the newly-selected language,
as much as possible.

This was done across the board with two commands:

$ find . -iname \*.html -exec sed -i 's!<button name="language" value="he">עברית</button>!<a href="/2018/he" class="nav-link">עברית</a>!g' {} \;
$ find . -iname \*.html -exec sed -i 's!<button name="language" value="en">English</button>!<a href="/2018/" class="nav-link">English</a>!g' {} \;

A small thing is lost in this simplistic translation: Originally,
a language-change from a page which exists in both languages would
take you to the other-language-version of the same page. Now, it
always goes to the home page in the other language. I think keeping
this behavior, at this point, is not worth the trouble it would take.