Fix paging links in blog
In the Django-based site, those are `href="?page=2"` etc. links.
When wget mirrors, the pages are stored as files named
`index.html?page=2.html` etc. These are names that don't fit
the links, and are difficult to get in a URL anyway.

The fix is in two steps. First, rename the files, replacing
the `?` with a `@`:

$ for f in $(find blog -iname \*page\*); do git mv "$f" "${f/\?/@}"; done

Then, fix the links accordingly. It turns out that out of the page
files themselves, these links appeared only in two files, so:

$ sed -i 's/\?page=\(.\)/index.html@page=\1.html/g' blog/index.html
$ sed -i 's/\?page=\(.\)/index.html@page=\1.html/g' blog/section/sponsors/index.html
$ for f in $(find blog -iname \*page\*); do sed -i 's/\?page=\(.\)/index.html@page=\1.html/g' $f; done

Some explanation: `page=\(.\)` is a regex that matches `page=x` for
whatever `x` is there, and the use of the parentheses `\(` and '\)'
means that the `x` is "captured", and can then be referenced using
`\1`. So this turns `?page=x` to `index.html@page=x.html`, for any
character `x` (the only relevent ones are `1` and `2`, but...).